refactor(epic2): C2 Swap & Strangle
