+++ shell/src/app/chat/chat-coordinator/chat-coordinator.spec.ts
@@ -19,7 +19,7 @@ import {TestBed} from '@angular/core/testing';
 import {signal} from '@angular/core';
 import {describe, it, expect, beforeEach, vi} from 'vitest';
 import {ChatCoordinator} from './chat-coordinator';
-import {redactApiKey} from './error-utils';
+import {redactApiKey} from '../chat-service/error-utils';
 import {CatalogManagement} from '../../storage/catalog-management/catalog-management';
 import {Catalog} from '../../storage/models/catalog-storage.model';
 import {ChatState, LlmLogEntry, LlmLogType} from '../chat-state/chat-state';
