--- shell/src/app/chat/chat-service/chat-coordinator.ts
@@ -1,1136 +1,0 @@
-/**
- * Copyright 2026 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import {computed, effect, inject, Injectable, signal, untracked} from '@angular/core';
-import {A2uiComponentInstance, PreviewBridgeMessageType, RenderA2uiItem} from 'a2ui-bridge';
-import {COMMON_TYPES_SCHEMA} from '../../gallery/schema/common-types-schema';
-import {AppConfigProvider} from '../../settings/app-config-provider/app-config-provider';
-import {CrossFrameValidator} from '../../shell/cross-frame-validator/cross-frame-validator';
-import {CatalogManagement} from '../../storage/catalog-management/catalog-management';
-import {PromptTurnType, UsageTrackingService} from '../../usage-tracking/usage-tracking.service';
-import {formatJson, tryParseJsonArray} from '../../utils/json';
-import {ChatState, LlmLogType} from '../chat-state/chat-state';
-import {
-  Attachment,
-  CANCEL_ERROR_NAME,
-  LlmClient,
-  LlmMessage,
-  LlmStreamResponse,
-  MessageRole,
-} from '../llm-client/llm-client';
-import {PipelineStatus} from '../pipeline-status/pipeline-status';
-import {StateSync} from '../state-sync/state-sync';
-import {ChatCleaner} from './chat-cleaner';
-import {cleanErrorMessage, redactApiKey} from './error-utils';
-
-@Injectable({
-  providedIn: 'root',
-})
-/**
- * Dynamic chat panel coordinator managing system prompt generation using
- * dynamic component configurations. Manages LLM completions transport
- * streams, self-healing parsers, schemas typo corrections, and gateway
- * error fallbacks.
- */
-export class ChatCoordinator {
-  private readonly catalogManagement = inject(CatalogManagement);
-  private readonly configProvider = inject(AppConfigProvider);
-  private readonly stateSync = inject(StateSync);
-  private readonly chatState = inject(ChatState);
-  private readonly llmClient = inject(LlmClient);
-  private readonly chatCleaner = inject(ChatCleaner);
-  private readonly usageTrackingService = inject(UsageTrackingService);
-
-  /** Reactively mapped rendering pipeline execution milestones. */
-  readonly pipelineStatus = this.chatState.pipelineStatus;
-
-  /**
-   * Public programmatic lock signal protecting against typing deadlocks
-   * during streams.
-   */
-  readonly isProgrammaticStreamActive = this.chatState.isProgrammaticStreamActive;
-
-  /** Turn index counter for telemetry. */
-  readonly currentTurnIndex = signal(0);
-
-  private activePromptId: string | null = null;
-  private lastSeenRendererUrl = '';
-  private isFirstUrlEffectRun = true;
-
-  constructor() {
-    // Effect monitoring dynamic host preview configurations mapping cache
-    // resets
-    effect(() => {
-      const url = this.configProvider.rendererUrl();
-      untracked(() => {
-        if (this.isFirstUrlEffectRun) {
-          this.isFirstUrlEffectRun = false;
-          this.lastSeenRendererUrl = url;
-          return;
-        }
-        if (this.lastSeenRendererUrl !== url) {
-          queueMicrotask(() => this.wipeEnvironmentCache());
-        }
-        this.lastSeenRendererUrl = url;
-      });
-    });
-  }
-
-  /**
-   * Resets turns logs history, overlays milestones, and locks indicators.
-   */
-  wipeEnvironmentCache(): void {
-    this.currentTurnIndex.set(0);
-    this.activePromptId = null;
-    this.chatState.setChatHistory([]);
-    this.chatState.setPipelineStatus(PipelineStatus.IDLE);
-    this.chatState.setProgrammaticStreamActive(false);
-    this.chatState.clearRawLlmHistory();
-    this.stateSync.flushDraft();
-  }
-
-  /**
-   * Constructs System instructions prepended message logs context.
-   */
-  getFullMessageContext(): LlmMessage[] {
-    return [
-      {
-        role: MessageRole.SYSTEM,
-        content: this.systemPrompt(),
-      },
-      ...this.chatState.chatHistory().filter(m => m.role !== MessageRole.ERROR),
-    ];
-  }
-
-  private activeStreamResponse?: LlmStreamResponse;
-  private isCancelRequested = false;
-
-  /**
-   * Cancels the currently active streaming request if there is one.
-   */
-  cancelActiveStream(): void {
-    this.isCancelRequested = true;
-    if (this.activePromptId) {
-      this.usageTrackingService.trackChatCancel({
-        promptId: this.activePromptId,
-        turnIndex: this.currentTurnIndex(),
-        pipelineStatus: this.pipelineStatus(),
-      });
-    }
-    if (this.activeStreamResponse && this.activeStreamResponse.cancel) {
-      this.activeStreamResponse.cancel();
-    }
-  }
-
-  private emitPromptTracking(
-    prompt: string,
-    attachments: Attachment[],
-    options?: {promptId?: string; promptTurnIndex?: number; retryOfPromptId?: string},
-  ): string {
-    const isRetry = !!options?.retryOfPromptId;
-    const promptTurnIndex = options?.promptTurnIndex ?? this.currentTurnIndex() + 1;
-    this.currentTurnIndex.set(promptTurnIndex);
-
-    const catalogObj = this.catalogManagement.activeCatalog();
-    const catalogId = catalogObj ? catalogObj.catalogId || catalogObj.$id || '' : '';
-
-    const hasScreenshot = attachments.some(
-      a => a.name === 'screenshot.png' || a.mimeType?.startsWith('image/'),
-    );
-    const nonScreenshotAttachments = attachments.filter(
-      a => a.name !== 'screenshot.png' && !a.mimeType?.startsWith('image/'),
-    );
-
-    if (isRetry) {
-      return this.usageTrackingService.trackChatRetry({
-        promptId: options?.promptId,
-        catalogId,
-        turnIndex: promptTurnIndex,
-        attemptNumber: 2,
-        retryOfPromptId: options?.retryOfPromptId,
-      });
-    } else {
-      return this.usageTrackingService.trackChatPrompt({
-        promptId: options?.promptId,
-        catalogId,
-        turnType: promptTurnIndex === 1 ? PromptTurnType.INITIAL : PromptTurnType.FOLLOWUP,
-        turnIndex: promptTurnIndex,
-        attemptNumber: 1,
-        hasScreenshot,
-        attachmentCount: nonScreenshotAttachments.length,
-      });
-    }
-  }
-
-  /**
-   * Dispatches a fresh text instruction, triggers GenAI completions
-   * in-stream, buffers packets, runs auto-repair healing and schema
-   * validation blocks.
-   */
-  async submitPrompt(
-    prompt: string,
-    attachments: Attachment[] = [],
-    options?: {promptId?: string; promptTurnIndex?: number; retryOfPromptId?: string},
-  ): Promise<void> {
-    if (this.chatState.isProgrammaticStreamActive()) {
-      console.warn('[ChatCoordinator] Blocked submitPrompt: programmatic stream is active.');
-      return;
-    }
-    const trimmed = prompt.trim();
-    if (!trimmed && attachments.length === 0) return;
-
-    const promptId = this.emitPromptTracking(trimmed, attachments, options);
-    this.activePromptId = promptId;
-
-    // Lock UI controls and transition state indicators to receiving stream
-    this.chatState.setProgrammaticStreamActive(true);
-    this.chatState.setPipelineStatus(PipelineStatus.RECEIVING_STREAM);
-
-    // Append user prompt text turn to chat history
-    this.chatState.updateChatHistory(h => [
-      ...h,
-      {
-        role: MessageRole.USER,
-        content: trimmed,
-        attachments: attachments.length > 0 ? attachments : undefined,
-        promptId,
-      },
-    ]);
-
-    // Construct system-prepended context matching conversational bounds
-    const fullContext = this.getFullMessageContext();
-
-    // Log the raw LLM request telemetry
-    this.chatState.addRawLlmLog(LlmLogType.REQUEST, fullContext);
-
-    // Push initial model turn placeholder with loading pulse indicator to history
-    this.chatState.updateChatHistory(h => [
-      ...h,
-      {
-        role: MessageRole.MODEL,
-        content: this.chatCleaner.appendPulse(''),
-      },
-    ]);
-
-    try {
-      this.isCancelRequested = false;
-      // Trigger streaming GenAI completions call using client facade
-      const responseStream = await this.llmClient.chatStream(fullContext);
-
-      // If a cancel was requested while the stream connection was establishing
-      if (this.isCancelRequested) {
-        if (responseStream.cancel) responseStream.cancel();
-        const err = new Error('Cancelled');
-        err.name = CANCEL_ERROR_NAME;
-        throw err;
-      }
-
-      this.activeStreamResponse = responseStream;
-
-      // Loop asynchronously over incoming stream packets to compile text
-      let accumulatedRawText = '';
-      let accumulatedThinking = '';
-      for await (const chunk of responseStream.contentStream) {
-        accumulatedRawText += chunk.content;
-        if (chunk.thinking) {
-          accumulatedThinking += chunk.thinking;
-        }
-
-        // Update history bubble in real-time with trailing pulse indicator
-        this.chatState.updateChatHistory(history => {
-          const updated = [...history];
-          const lastIdx = updated.length - 1;
-          if (updated[lastIdx]?.role === MessageRole.MODEL) {
-            updated[lastIdx] = {
-              role: MessageRole.MODEL,
-              content: this.chatCleaner.appendPulse(accumulatedRawText),
-              thinking: accumulatedThinking,
-            };
-          }
-          return updated;
-        });
-      }
-
-      // Stream exhausted, resolve final complete text and remove visual loading indicator
-      const finalRawText = await responseStream.complete;
-
-      // Log the raw LLM response telemetry
-      this.chatState.addRawLlmLog(LlmLogType.RESPONSE, finalRawText);
-      this.chatState.updateChatHistory(history => {
-        const updated = [...history];
-        const lastIdx = updated.length - 1;
-        if (updated[lastIdx]?.role === MessageRole.MODEL) {
-          updated[lastIdx] = {
-            role: MessageRole.MODEL,
-            content: finalRawText,
-            thinking: accumulatedThinking,
-          };
-        }
-        return updated;
-      });
-
-      this.chatState.setPipelineStatus(PipelineStatus.RECEIVED_RAW);
-      await this.processRawLlmPayload(finalRawText);
-    } catch (err: unknown) {
-      // If it was cancelled, don't show an error. Just leave what was generated or remove the bubble.
-      // But we probably want to just reset the UI lock.
-      if (err && typeof err === 'object' && 'name' in err && err.name === CANCEL_ERROR_NAME) {
-        this.chatState.setPipelineStatus(PipelineStatus.IDLE);
-        this.chatState.setProgrammaticStreamActive(false);
-        // Replace trailing pulse or partial JSON with stopped message, and force non-snapshot
-        this.chatState.updateChatHistory(history => {
-          const updated = [...history];
-          const lastIdx = updated.length - 1;
-          if (updated[lastIdx]?.role === MessageRole.MODEL) {
-            updated[lastIdx] = {
-              ...updated[lastIdx],
-              content: '*You stopped this response.*',
-            };
-          }
-          return updated;
-        });
-      } else {
-        this.handleConnectivityError(err, trimmed, attachments, promptId);
-      }
-    } finally {
-      this.activeStreamResponse = undefined;
-    }
-  }
-
-  /**
-   * Post-processes, extracts, syntax heals, and validates raw JSON lines.
-   */
-  private async processRawLlmPayload(rawText: string): Promise<void> {
-    // Stage 1: Parse and Syntax Healing
-    let parsedBlocks: unknown[] = [];
-    try {
-      parsedBlocks = this.parseAndHealJsonLines(rawText);
-    } catch (err: unknown) {
-      this.chatState.setPipelineStatus(PipelineStatus.FAILED);
-      this.chatState.setProgrammaticStreamActive(false);
-      throw err;
-    }
-
-    // Stage 2: Schema Validation
-    this.chatState.setPipelineStatus(PipelineStatus.VALIDATING);
-    try {
-      // Verify basic schema envelopes using pre-existing CrossFrameValidator
-      const mockEnvMsg = {
-        type: PreviewBridgeMessageType.RENDER_A2UI,
-        payload: parsedBlocks,
-      };
-
-      const validationErrors: string[] = [];
-      const isValidEnvelope = CrossFrameValidator.validateOutgoingMessage(
-        mockEnvMsg,
-        validationErrors,
-      );
-
-      if (!isValidEnvelope) {
-        throw new Error(
-          `Outgoing message envelope validation failed:\n${validationErrors.join('\n')}`,
-        );
-      }
-
-      // Catalog Component Schema Check & Name Typos Healing
-      this.runCatalogComponentSchemaCheck(parsedBlocks);
-
-      // Stage 3: Ready & Commit Layout Wipes
-      this.chatState.setPipelineStatus(PipelineStatus.READY);
-
-      // Turn list of updates back into raw formatted JSON text to write to
-      // editor draft
-      const finalLayoutText = formatJson(parsedBlocks);
-
-      // Update latest MODEL turn message in history to normalized JSON string
-      this.chatState.updateChatHistory(history => {
-        const updated = [...history];
-        const lastIdx = updated.length - 1;
-        if (updated[lastIdx]?.role === MessageRole.MODEL) {
-          updated[lastIdx] = {
-            ...updated[lastIdx],
-            content: finalLayoutText,
-          };
-        }
-        return updated;
-      });
-
-      // Commit layout synchronously to editor viewport before releasing
-      // lockout
-      this.stateSync.commitLayoutFromLlm(finalLayoutText);
-
-      // Release panel textareas lockout synchronously to avoid race
-      // condition escapes
-      this.chatState.setProgrammaticStreamActive(false);
-    } catch (err: unknown) {
-      this.chatState.setPipelineStatus(PipelineStatus.FAILED);
-      this.chatState.setProgrammaticStreamActive(false);
-      throw err;
-    }
-  }
-
-  /**
-   * Robust parser extracting JSON objects from blocks, performing syntax
-   * repairs.
-   */
-  private parseAndHealJsonLines(text: string): unknown[] {
-    if (this.chatCleaner.extractCodeFences(text).hasFences) {
-      this.chatState.setPipelineStatus(PipelineStatus.HEALING);
-    }
-    const content = this.chatCleaner.cleanPayload(text);
-
-    // Attempt full JSON parsing before line-by-line processing
-    const parsedArray = tryParseJsonArray(content);
-    if (parsedArray) {
-      return parsedArray;
-    }
-    try {
-      const parsedSingle = JSON.parse(content);
-      if (Array.isArray(parsedSingle)) {
-        return parsedSingle;
-      }
-      if (parsedSingle && typeof parsedSingle === 'object') {
-        return [parsedSingle];
-      }
-    } catch {
-      // Continue to line-by-line healing
-    }
-
-    const lines = content
-      .split('\n')
-      .map(l => l.trim())
-      .filter(l => l.length > 0);
-    const parsedBlocks: unknown[] = [];
-
-    for (const line of lines) {
-      // Skip Markdown code tags if they leaked, or general prompt filler
-      // text lines
-      if (line.startsWith('```') || (!line.startsWith('{') && !line.startsWith('['))) {
-        continue;
-      }
-
-      try {
-        parsedBlocks.push(JSON.parse(line));
-      } catch (err) {
-        // Syntax Healing Loop
-        this.chatState.setPipelineStatus(PipelineStatus.HEALING);
-        const healedObj = this.attemptSyntaxHealing(line);
-        if (healedObj !== null) {
-          parsedBlocks.push(healedObj);
-        } else {
-          // If it looks like A2UI JSON but couldn't be repaired, throw
-          // validation error
-          if (line.includes('"version"') || line.includes('"createSurface"')) {
-            throw new Error(`Syntax recovery failed for corrupted JSON Line:\n"${line}"`);
-          }
-        }
-      }
-    }
-
-    if (parsedBlocks.length === 0) {
-      throw new Error('No valid A2UI JSON layout command block could be parsed or recovered.');
-    }
-
-    return parsedBlocks;
-  }
-
-  /**
-   * Attempts structural syntax patching on broken JSON strings.
-   */
-  private attemptSyntaxHealing(line: string): unknown | null {
-    let patched = line.trim();
-
-    // Repair 1: Strip trailing commas inside properties arrays
-    patched = patched.replace(/,\s*([\]}])/g, '$1');
-
-    // Repair 2: Auto-close braces
-    try {
-      return JSON.parse(patched);
-    } catch (e) {
-      // Loop to try appending up to 5 missing closing curly braces
-      for (let i = 1; i <= 5; i++) {
-        try {
-          return JSON.parse(patched + '}'.repeat(i));
-        } catch (_) {}
-      }
-
-      // Loop to try appending matching square brackets then curly braces
-      for (let i = 1; i <= 3; i++) {
-        for (let j = 1; j <= 3; j++) {
-          try {
-            return JSON.parse(patched + ']'.repeat(i) + '}'.repeat(j));
-          } catch (_) {}
-        }
-      }
-    }
-
-    return null;
-  }
-
-  /**
-   * Validates parsed components against custom catalog schemas, healing name
-   * typos, mapping legacy names, and recursively stripping out custom mock
-   * rules configurations.
-   */
-  private runCatalogComponentSchemaCheck(parsedBlocks: unknown[]): void {
-    const catalog = this.catalogManagement.activeCatalog();
-    const componentsObj = catalog?.components;
-    const componentHealMap: Record<string, string> = {};
-
-    if (componentsObj) {
-      for (const key of Object.keys(componentsObj)) {
-        const normalizedKey = key.toLowerCase().replace(/[^a-z]/g, '');
-        componentHealMap[normalizedKey] = key;
-      }
-    }
-
-    const SYNONYM_MAP: Record<string, string> = {
-      textbox: 'textfield',
-      textinput: 'textfield',
-      rowlayout: 'row',
-      columnlayout: 'column',
-      choice: 'choicepicker',
-      datepicker: 'datetimeinput',
-      datetimepicker: 'datetimeinput',
-    };
-
-    for (const block of parsedBlocks) {
-      if (!block || typeof block !== 'object') {
-        continue;
-      }
-      const bObj = block as RenderA2uiItem;
-      const updateComponents = bObj.updateComponents;
-      if (
-        !updateComponents ||
-        typeof updateComponents !== 'object' ||
-        !Array.isArray(updateComponents.components)
-      ) {
-        continue;
-      }
-
-      const cleanedComponents: unknown[] = [];
-      for (const comp of updateComponents.components) {
-        if (!comp || typeof comp !== 'object' || Array.isArray(comp)) {
-          cleanedComponents.push(comp);
-          continue;
-        }
-
-        const compObj = comp as A2uiComponentInstance;
-        let compType = compObj.component;
-
-        // legacy property "name" fallback: heal to "component" key mapping
-        if (compObj['name'] && !compObj.component) {
-          this.chatState.setPipelineStatus(PipelineStatus.HEALING);
-          compType = compObj['name'] as string;
-          compObj.component = compType;
-          delete compObj['name'];
-        }
-
-        if (typeof compType !== 'string') {
-          throw new Error('Component declaration is missing component type name string.');
-        }
-
-        let targetType = compType;
-
-        // Schema validation (only if catalog is actively loaded with components)
-        if (componentsObj) {
-          if (!componentsObj[compType]) {
-            // Unrecognized component type - check case-insensitive lookup!
-            const normalized = compType.toLowerCase().replace(/[^a-z]/g, '');
-            let healedType = componentHealMap[normalized];
-
-            // If not found directly, check synonym translation dictionary
-            if (!healedType) {
-              const synonymTarget = SYNONYM_MAP[normalized];
-              if (synonymTarget) {
-                healedType = componentHealMap[synonymTarget];
-              }
-            }
-
-            if (healedType && componentsObj[healedType]) {
-              this.chatState.setPipelineStatus(PipelineStatus.HEALING);
-              targetType = healedType;
-            } else {
-              // Fuzzy search matches
-              const fuzzyMatch = normalized
-                ? Object.keys(componentsObj).find(
-                    key =>
-                      key.toLowerCase().includes(normalized) ||
-                      normalized.includes(key.toLowerCase()),
-                  )
-                : undefined;
-
-              if (fuzzyMatch) {
-                this.chatState.setPipelineStatus(PipelineStatus.HEALING);
-                targetType = fuzzyMatch;
-              } else {
-                throw new Error(
-                  `Validation failure: Component type "${compType}" is ` +
-                    'not registered in the active custom catalog.',
-                );
-              }
-            }
-          }
-        }
-
-        // Recursively strip out dynamic mock setups configuration fields
-        const cleanedComp = this.sanitizeComponentObject(compObj);
-        // Restore corrected element name
-        cleanedComp.component = targetType;
-        cleanedComponents.push(cleanedComp);
-      }
-
-      // Commit sanitized array back in-place
-      updateComponents.components = cleanedComponents;
-    }
-  }
-
-  /**
-   * Unifies recursive sanitization traversal and strips out dynamic mock setups
-   * configurations recursively.
-   */
-  private sanitizeValue(val: unknown): unknown {
-    if (val === null || typeof val !== 'object') {
-      return val;
-    }
-
-    if (Array.isArray(val)) {
-      return val.map(item => this.sanitizeValue(item));
-    }
-
-    const obj = val as Record<string, unknown>;
-    const cleaned: Record<string, unknown> = {};
-
-    for (const [key, propVal] of Object.entries(obj)) {
-      if (key === 'rules' || /^mock/i.test(key)) {
-        continue;
-      }
-      cleaned[key] = this.sanitizeValue(propVal);
-    }
-
-    return cleaned;
-  }
-
-  /**
-   * Recursively sanitizes component declarations maps.
-   * Strips out dynamic rules configs matching /rules/ or prefix /^mock/i.
-   */
-  private sanitizeComponentObject(obj: A2uiComponentInstance): A2uiComponentInstance {
-    return this.sanitizeValue(obj) as A2uiComponentInstance;
-  }
-
-  readonly TEST_ONLY = {
-    sanitizeComponentObject: (obj: A2uiComponentInstance) => this.sanitizeComponentObject(obj),
-  };
-
-  /**
-   * Connectivity Exception Handling: bubbles diagnostics stack details.
-   * Instantly dismisses overlays locks on network, proxy, or auth failures
-   * to restore workspace editing controls immediately.
-   */
-  private isConnectivityError(lowerMsg: string): boolean {
-    return (
-      lowerMsg.includes('failed to fetch') ||
-      lowerMsg.includes('fetch') ||
-      lowerMsg.includes('timeout') ||
-      lowerMsg.includes('504') ||
-      lowerMsg.includes('proxy') ||
-      lowerMsg.includes('networkerror') ||
-      lowerMsg.includes('connection') ||
-      lowerMsg.includes('401') ||
-      lowerMsg.includes('403') ||
-      lowerMsg.includes('credential') ||
-      lowerMsg.includes('quota') ||
-      lowerMsg.includes('blocked') ||
-      lowerMsg.includes('503') ||
-      lowerMsg.includes('unavailable') ||
-      lowerMsg.includes('api key') ||
-      lowerMsg.includes('apikey')
-    );
-  }
-
-  private parseError(
-    lowerMsg: string,
-    cleanMsg: string,
-    originalPrompt?: string,
-  ): {
-    errorTitle: string;
-    errorMessage: string;
-    errorTip: string;
-    isRetryable: boolean;
-    showDetails: boolean;
-    errorDetails?: string;
-  } {
-    // Default values (Connectivity Failure)
-    const errorTitle = 'Connectivity Failure';
-    const isJson = cleanMsg.trim().startsWith('{');
-    const errorMessage = isJson ? 'A connectivity error occurred.' : cleanMsg;
-    const errorDetails = isJson ? 'Details: ' + cleanMsg : undefined;
-    const errorTip =
-      'Tip: Please check your network proxy configurations or verify your settings to restore connections.';
-    const isRetryable = !!originalPrompt;
-    const showDetails = true;
-
-    const isValidationError =
-      lowerMsg.includes('validation') ||
-      lowerMsg.includes('syntax recovery') ||
-      lowerMsg.includes('validation failure');
-
-    if (isValidationError) {
-      return {
-        errorTitle: 'Validation Failure',
-        errorMessage: 'The generated layout contains invalid components or structure.',
-        errorTip:
-          'Tip: Try rephrasing your prompt to guide the model to generate valid components.',
-        isRetryable: !!originalPrompt,
-        showDetails: true,
-        errorDetails: 'Details: ' + cleanMsg,
-      };
-    }
-
-    if (lowerMsg.includes('503') || lowerMsg.includes('unavailable')) {
-      return {
-        errorTitle: 'Service Unavailable',
-        errorMessage: 'The generative service is temporarily unavailable. Please try again later.',
-        errorTip: '',
-        isRetryable: true,
-        showDetails: false,
-      };
-    }
-
-    if (lowerMsg.includes('high demand')) {
-      return {
-        errorTitle: 'Model High Demand',
-        errorMessage:
-          'This model is currently experiencing high demand. Spikes in demand are usually temporary. Please try again later.',
-        errorTip: '',
-        isRetryable: true,
-        showDetails: false,
-      };
-    }
-
-    if (lowerMsg.includes('timeout') || lowerMsg.includes('504')) {
-      return {
-        errorTitle: 'REST Gateway Timeout',
-        errorMessage: 'Remote generation service did not respond.',
-        errorDetails: 'Details: ' + cleanMsg,
-        errorTip,
-        isRetryable,
-        showDetails: true,
-      };
-    }
-
-    if (lowerMsg.includes('api key') || lowerMsg.includes('apikey')) {
-      return {
-        errorTitle: 'Invalid API Key',
-        errorMessage: 'The provided Gemini API key is invalid or missing.',
-        errorDetails: 'Details: ' + cleanMsg,
-        errorTip:
-          'Tip: Please update your third-party Gemini developer API key on the settings page to restore connections.',
-        isRetryable,
-        showDetails: true,
-      };
-    }
-
-    if (
-      lowerMsg.includes('auth') ||
-      lowerMsg.includes('401') ||
-      lowerMsg.includes('403') ||
-      lowerMsg.includes('credential')
-    ) {
-      return {
-        errorTitle: 'Authentication Refused',
-        errorMessage: 'Authentication failed. Please verify your credentials in Settings.',
-        errorDetails: 'Details: ' + cleanMsg,
-        errorTip,
-        isRetryable,
-        showDetails: true,
-      };
-    }
-
-    if (lowerMsg.includes('quota') || lowerMsg.includes('blocked') || lowerMsg.includes('429')) {
-      return {
-        errorTitle: 'GenAI Service Blocked',
-        errorMessage: 'Resource quota depleted or content safety limits triggered.',
-        errorDetails: 'Details: ' + cleanMsg,
-        errorTip,
-        isRetryable,
-        showDetails: true,
-      };
-    }
-
-    return {
-      errorTitle,
-      errorMessage,
-      errorTip,
-      isRetryable,
-      showDetails,
-      errorDetails,
-    };
-  }
-
-  private handleConnectivityError(
-    err: unknown,
-    originalPrompt?: string,
-    attachments: Attachment[] = [],
-    promptId?: string,
-  ): void {
-    const rawError = err instanceof Error ? err.message : String(err);
-    const lowerMsg = rawError.toLowerCase();
-    const cleanMsg = cleanErrorMessage(rawError);
-
-    if (this.isConnectivityError(lowerMsg)) {
-      this.chatState.setPipelineStatus(PipelineStatus.IDLE);
-    } else {
-      this.chatState.setPipelineStatus(PipelineStatus.FAILED);
-    }
-    this.chatState.setProgrammaticStreamActive(false);
-
-    const parsed = this.parseError(lowerMsg, cleanMsg, originalPrompt);
-
-    let exceptionDetails = '';
-    if (err instanceof Error) {
-      exceptionDetails = 'Exception: ' + err.message + '\nStack: ' + (err.stack || 'None');
-    } else {
-      exceptionDetails = 'Unknown Exception: ' + JSON.stringify(err);
-    }
-
-    let combinedDetails = '';
-    if (parsed.errorDetails) {
-      combinedDetails += parsed.errorDetails + '\n\n';
-    }
-    combinedDetails += exceptionDetails;
-
-    // Apply API key redaction
-    const redactedErrorMessage = redactApiKey(parsed.errorMessage);
-    const redactedErrorDetails = parsed.showDetails ? redactApiKey(combinedDetails) : undefined;
-    const redactedErrorTip = parsed.showDetails ? redactApiKey(parsed.errorTip) : undefined;
-
-    console.error('Gemini chat execution failed:', err);
-
-    this.chatState.updateChatHistory(history => {
-      const updated = [...history];
-      const lastIdx = updated.length - 1;
-      const errorBubble = {
-        role: MessageRole.ERROR,
-        content: redactedErrorMessage,
-        errorTitle: parsed.errorTitle,
-        errorMessage: redactedErrorMessage,
-        errorDetails: redactedErrorDetails,
-        errorTip: redactedErrorTip,
-        promptId,
-        ...(parsed.isRetryable ? {isRetryable: true, originalPrompt, attachments} : {}),
-      };
-      if (lastIdx >= 0 && updated[lastIdx].role === MessageRole.MODEL) {
-        updated[lastIdx] = errorBubble;
-        return updated;
-      }
-      updated.push(errorBubble);
-      return updated;
-    });
-  }
-
-  /**
-   * A dynamic, reactive, computed signal property constructing conformed JSON
-   * catalog schema specifications system instructions.
-   */
-  readonly systemPrompt = computed<string>(() => {
-    const catalog = this.catalogManagement.activeCatalog();
-    if (!catalog) {
-      return `
-  # A2UI Generation Expert
-
-  ## Role
-  You are an expert A2UI generation assistant. Your role is to translate user
-  requests—whether provided as text instructions, UI wireframes, screenshots,
-  or mockup images—into valid A2UI v0.9 interactive user interfaces.
-      `;
-    }
-
-    return this.generateSystemPrompt(formatJson(catalog));
-  });
-
-  private generateSystemPrompt(catalog: string): string {
-    return `
-  # A2UI Generation Expert
-
-  ## Role
-  You are an expert A2UI generation assistant. Your role is to translate user
-  requests—whether provided as text instructions, UI wireframes, screenshots,
-  or mockup images—into valid A2UI v0.9 interactive user interfaces.
-
-  ## Catalog Allowlist & Component Rules
-
-  You MUST strictly enforce the following rules regarding component selection
-  and schema compliance:
-  1. **Strict Component Allowlist**: You MUST use ONLY the component types
-     defined as keys in the "components" map of the active catalog schema
-     provided below.
-  2. **No Hallucinated Component Names**: Never invent, guess, or mix
-     component names from other libraries or catalogs. For example, if only
-     "Column", "Row", "Text", and "Button" are present in the active catalog
-     schema below, emitting "MaterialColumn", "MaterialText", or "Div" is
-     strictly INVALID.
-  3. **No Hallucinated Properties**: Include ONLY properties explicitly
-     defined in the JSON Schema for that specific component type in the
-     active catalog. Do NOT emit unauthorized keys (e.g., "rules", "mock*",
-     or unsupported CSS/styling parameters).
-
-  ### Active Catalog Schema (Mandatory Allowlist)
-  \`\`\`json
-  ${catalog}
-  \`\`\`
-
-  ### Common Schema Types
-  Common structural types referenced by $ref in the catalog schema (e.g.,
-  DataBinding, Action, Event, DynamicString, etc.) are defined here:
-  \`\`\`json
-  ${formatJson(COMMON_TYPES_SCHEMA)}
-  \`\`\`
-
-  ## Output Format: Strict A2UI JSON Lines (JSONL)
-
-  Your output MUST be valid **A2UI JSON Lines (JSONL)**:
-  1. **One JSON Object Per Line**: Each A2UI message MUST be formatted as a
-     single, valid JSON object on its own line, terminated by a newline
-     character (\\n). Do NOT pretty-print or split a single JSON object across
-     multiple lines.
-  2. **Required Version & Command**: Every message object MUST include
-     "version": "v0.9" at the top level and specify exactly one A2UI
-     command: "createSurface", "updateComponents", "updateDataModel", or
-     "deleteSurface".
-  3. **No Markdown or Preamble**: Output ONLY raw JSON Lines. Do NOT wrap
-     your response in markdown code fences (such as \`\`\`jsonl or \`\`\`). Do
-     NOT include any conversational text, greetings, explanations,
-     scratchpad analysis, or summary before or after the JSON Lines.
-  4. **Direct Parseability**: Every line in your response MUST be
-     independently parseable by JSON.parse().
-
-  ## Multimodal & Image-to-UI Guidelines
-
-  When an image, wireframe, mockup, or UI screenshot is provided by the
-  user, adhere strictly to these visual translation principles:
-
-  ### 1. Visual Layout, Scope & Sizing Fidelity
-  * **Root Container Bounding**: The root component ("id": "root") MUST match
-    the visual boundary of the primary UI card, form, or dialog shown. Do
-    NOT extract ambient background titles, file names, or browser canvas
-    headers outside the visual card boundary unless explicitly requested.
-  * **Flex Orientation Mapping**:
-    - Elements arranged top-to-bottom MUST map to vertical layout containers
-      defined in the active catalog (e.g., Column).
-    - Elements arranged left-to-right MUST map to horizontal layout
-      containers defined in the active catalog (e.g., Row).
-  * **Full-Width Stretch Mandate**: When an element (such as a primary CTA
-    button, input field, or card) visually spans the full width of its
-    parent container in the screenshot, configure its layout/alignment
-    properties to stretch full-width (e.g., setting "align": "stretch" on
-    the parent container or applying full-width properties supported by
-    the active catalog) rather than rendering as a compact inline element.
-  * **Container Spacing & Clipping Prevention**: Ensure root layout
-    containers (Column) and nested sections maintain proper vertical
-    padding, spacing, and scrollability so that bottom elements (such as
-    footer actions or trailing list items) are never cut off or clipped.
-  * **No Unseen Separators Rule**: Do NOT insert "Divider" lines or border
-    components unless a distinct horizontal or vertical line separator is
-    literally visible in the screenshot.
-  * **Visual Reading Order**: List child IDs in children arrays in strict
-    visual reading order (top-to-bottom, left-to-right).
-
-  ### 2. Catalog-Aware Component Mapping
-  Map visual elements to the most specific matching component type from the
-  "components" allowlist of the active catalog schema provided above:
-  * **Headings & Titles** -> Text component with heading typography styles
-    (usageHint: "h1" | "h2" | "h3" or equivalent variant property in the
-    active catalog schema).
-  * **Body Text & Captions** -> Text component with body or caption
-    typography styles (usageHint: "body" | "caption").
-  * **Interactive Buttons** -> Button/IconButton component in the active
-    catalog schema. Reflect visual prominence (e.g., primary filled vs.
-    secondary borderless/outlined) and preserve full-width intent.
-  * **Form Controls & Inputs** -> Text entry, date picker, selection/picker,
-    or toggle components defined in the active catalog schema.
-  * **Content Panels & Containers** -> Card, panel, or layout container
-    components defined in the active catalog schema wrapping child elements.
-  * **Repeated Lists & Collections** -> Layout container components with
-    dynamic item template declarations
-    (children: { "componentId": "...", "path": "/..." }).
-  * **CRITICAL**: Every generated "component" value MUST be an exact key
-    from the "components" map in the active catalog schema provided above.
-    Never invent or guess component names not present in the active catalog.
-
-  ### 3. Icon Fidelity, Custom SVG & Styling Intent
-  * **Composite Icon & Feature Matching**: Closely examine visual icon
-    shapes for composite features (e.g., a document with an edit badge, a
-    search icon with a filter indicator, or a custom symbol). First check if
-    the active catalog's Icon component includes an exact visual match in
-    its enum.
-  * **Custom SVG Fallback (No Close Icon Match)**: If an icon in the
-    screenshot has distinct visual features that do NOT have a close match
-    in the active catalog's predefined icon list:
-    - **Do NOT** substitute a visually mismatched, generic, or oversimplified
-      placeholder icon.
-    - **Fallback to SVG**: Generate an inline vector graphic instead using
-      one of the mechanisms supported by the active catalog schema:
-      1. If the Icon component in the active catalog accepts custom path
-         data, specify the svgPath property with a valid SVG path d string.
-      2. If an Image component is available in the active catalog, supply an
-         inline SVG Data URL in its url/image source property
-         ("data:image/svg+xml;utf8,<svg ...>...</svg>").
-  * **Visual Hierarchy**: Preserve typography scale, text weight, button
-    prominence, and color intent using supported catalog properties.
-
-  ### 4. Visual Affordance Recognition
-  Recognize common UI visual affordance symbols and map them strictly using
-  components defined in the active catalog schema provided above:
-  * **Downward Chevrons / Disclosure Carets (Collapsible Rows)**:
-    - **Visual Indicator**: Downward-facing arrows (∨, expand_more) at row
-      edges denote expandable/collapsible sections.
-    - **Catalog Mapping**: If the active catalog schema includes an expansion
-      or accordion component, use it. Otherwise, compose the row using
-      layout primitives in the catalog: e.g., a horizontal layout container
-      (Row) holding leading text/icons and a trailing downward icon.
-  * **Search Cues (Search Inputs)**:
-    - **Visual Indicator**: Magnifying glass symbols (🔍) inside or adjacent
-      to text entry boxes.
-    - **Catalog Mapping**: If a search component exists in the active catalog
-      schema, use it; otherwise, use a text input component paired with a
-      search icon.
-  * **Toggle Track & Thumb (Switches & Toggles)**:
-    - **Visual Indicator**: Pill-shaped track with a circular thumb (⚪━━).
-    - **Catalog Mapping**: Use a toggle, switch, or selection control
-      component defined in the active catalog schema.
-  * **Selection Controls (Option Pickers)**:
-    - **Visual Indicator**: Radio circles (◯ / 🔘), checkboxes (☐ / ☑), or
-      dropdown carets.
-    - **Catalog Mapping**: Look up selection, picker, or option components
-      in the active catalog schema; if none exist, compose using interactive
-      button components.
-  * **Pill Badges & Chips (Status & Tags)**:
-    - **Visual Indicator**: Small rounded rectangle or oval containing short
-      text/status labels.
-    - **Catalog Mapping**: Use a chip, badge, or tag component if defined in
-      the active catalog schema; otherwise, compose using a text component
-      inside a container or card.
-
-  ### 5. Grounding, Data Binding & Sequence
-  * **Complete Data Model Extraction**: ALL text strings, label names, image
-    URLs, options, and default values visible in the image MUST be extracted
-    into the updateDataModel payload.
-  * **JSON Pointer References**: Components in updateComponents MUST bind to
-    values in updateDataModel using valid JSON Pointers
-    (e.g., {"path": "/header/title"}). Do NOT hardcode visible text strings
-    inline when data binding is supported.
-  * **Strict Grounding**: Include ONLY visual elements present in the
-    screenshot. Do NOT hallucinate extra buttons, fields, or unrepresented
-    data streams.
-
-  ### 6. Image-to-UI Processing Sequence
-  When translating an image to A2UI, follow this internal mental sequence
-  (do NOT output any analysis or scratchpad text; output ONLY the final
-  JSONL messages):
-  1. **Analyze (Internal)**: Identify primary card boundaries, flex layout
-     directions, full-width element stretching, absence of unseen dividers,
-     container spacing, and composite icon details.
-  2. **Extract Data**: Extract all visible text strings, values, and list
-     items into updateDataModel.
-  3. **Build Component Tree**: Map visual elements strictly to active
-     catalog component types with exact icon names/SVGs, full-width
-     properties, and JSON Pointer paths.
-  4. **Emit JSONL Messages**: Output the single-line JSONL messages in
-     strict sequence (createSurface -> updateComponents -> updateDataModel).
-
-  ## Validation & Lifecycle Ordering
-
-  A complete A2UI payload consists of one or more message objects sent as
-  continuous JSON Lines. Every message object MUST include a top-level
-  "version": "v0.9" field.
-
-  The four primary messages you must use to manage a UI surface are:
-  1. **createSurface**: Sent **FIRST** to signal the client to create a new
-     surface. It defines the catalogId and optional theme parameters.
-  2. **updateComponents**: Used to define or update the UI component tree.
-     You must provide a flat list of components. One component MUST have an
-     id of "root".
-  3. **updateDataModel**: Used to define or update data values that the
-     components bind to.
-  4. **deleteSurface**: Signals the client to destroy the surface.
-
-  Typical sequence: createSurface -> updateComponents -> updateDataModel
-  (or combined/interleaved after creation).
-  When updating an existing UI in a multi-turn conversation, keep the
-  surfaceId consistent across turns.
-
-  ## Examples
-
-  **IMPORTANT**: The component names used in the examples below (Column, Text,
-  TextField, ChoicePicker, Button, etc.) are for structural illustration.
-  You MUST replace them with exact component names from the active catalog
-  schema provided above. In addition, code fences (\`\`\`jsonl) are shown
-  below for documentation readability only; do NOT include code fences in
-  your actual JSONL output.
-
-    * **Simple Example**: A basic column with text:
-      \`\`\`jsonl
-      {"version": "v0.9", "createSurface": {"surfaceId": "main", "catalogId": "https://a2ui.org/specification/v0_9/material_catalog.json"}}
-      {"version": "v0.9", "updateComponents": {"surfaceId": "main", "components": [{"id": "root", "component": "MaterialColumn", "children": ["header", "content"]}, {"id": "header", "component": "MaterialText", "text": "Welcome"}, {"id": "content", "component": "MaterialText", "text": {"path": "/message"}}]}}
-      {"version": "v0.9", "updateDataModel": {"surfaceId": "main", "path": "/message", "value": "Hello, world!"}}
-      \`\`\`
-
-    * **Complex Form Example**: A vacation booking form demonstrating advanced
-      Material form controls (\`MaterialDatepicker\`, \`MaterialSelect\`,
-      \`MaterialSlideToggle\`) and buttons using the modernized Material catalog:
-      \`\`\`jsonl
-      {"version": "v0.9", "createSurface": {"surfaceId": "vacation_booking", "catalogId": "https://a2ui.org/specification/v0_9/material_catalog.json"}}
-      {"version": "v0.9", "updateComponents": {"surfaceId": "vacation_booking", "components": [{"id": "root", "component": "MaterialColumn", "children": ["title", "destination_input", "checkin_datepicker", "checkout_datepicker", "room_type_select", "passenger_select", "flexible_dates_toggle", "search_button"]}, {"id": "title", "component": "MaterialText", "text": {"path": "/title_label"}, "usageHint": "h1"}, {"id": "destination_input", "component": "MaterialInput", "label": {"path": "/destination_label"}, "value": {"path": "/destination_value"}}, {"id": "checkin_datepicker", "component": "MaterialDatepicker", "label": {"path": "/checkin_label"}, "value": {"path": "/checkin_value"}}, {"id": "checkout_datepicker", "component": "MaterialDatepicker", "label": {"path": "/checkout_label"}, "value": {"path": "/checkout_value"}}, {"id": "room_type_select", "component": "MaterialSelect", "label": {"path": "/room_type_label"}, "value": {"path": "/room_type_value"}, "options": [{"label": "Standard Room", "value": "standard"}, {"label": "Deluxe Suite", "value": "deluxe"}]}, {"id": "passenger_select", "component": "MaterialSelect", "label": {"path": "/passenger_label"}, "value": {"path": "/passenger_value"}, "options": [{"label": "1 Passenger", "value": "1"}, {"label": "2 Passengers", "value": "2"}, {"label": "3+ Passengers", "value": "3"}]}, {"id": "flexible_dates_toggle", "component": "MaterialSlideToggle", "label": {"path": "/flexible_dates_label"}, "checked": {"path": "/flexible_dates_checked"}, "color": "primary"}, {"id": "search_button", "component": "MaterialButton", "label": {"path": "/search_label"}, "action": {"event": {"name": "searchVacation"}}}]}}
-      {"version": "v0.9", "updateDataModel": {"surfaceId": "vacation_booking", "value": {"title_label": "Book Your Dream Vacation", "destination_label": "Destination", "destination_value": "Hawaii", "checkin_label": "Check-in Date", "checkin_value": "2026-07-01", "checkout_label": "Check-out Date", "checkout_value": "2026-07-14", "room_type_label": "Room Type", "room_type_value": "standard", "passenger_label": "Passengers", "passenger_value": "2", "flexible_dates_label": "Flexible Dates (+/- 3 days)", "flexible_dates_checked": true, "search_label": "Search Flights & Hotels"}}}
-      \`\`\`
-
-    * **Dynamic List Example**: An example using templates to render a list of
-      items.
-      \`\`\`jsonl
-      {"version": "v0.9", "createSurface": {"surfaceId": "dynamic_list_demo", "catalogId": "https://a2ui.org/specification/v0_9/material_catalog.json"}}
-      {"version": "v0.9", "updateComponents": {"surfaceId": "dynamic_list_demo", "components": [{"id": "root", "component": "MaterialColumn", "children": ["title", "list_container"]}, {"id": "title", "component": "MaterialText", "text": "Dynamic List Demo"}, {"id": "list_container", "component": "MaterialColumn", "children": {"componentId": "item_template", "path": "/items"}}, {"id": "item_template", "component": "MaterialText", "text": {"path": "text"}}]}}
-      {"version": "v0.9", "updateDataModel": {"surfaceId": "dynamic_list_demo", "value": {"items": [{"text": "Item One"}, {"text": "Item Two"}]}}}
-      \`\`\`
-
-  ## Data Binding
-  Every component property value MUST come from the data model (with minor
-  exceptions for static primitives).
-  When referencing data in the data model, you MUST use valid JSON Pointer
-  syntax starting with /.
-
-  ## Actions and Context
-
-  When defining actions (e.g., on buttons), the \`context\` payload is a standard
-  JSON object, rather than an array of key-value pairs.
-
-  Example action definition:
-  \`\`\`json
-  "action": {
-    "event": {
-      "name": "selectItem",
-      "context": {
-        "itemId": "12345",
-        "itemName": {"path": "/selected/name"}
-      }
-    }
-  }
-  \`\`\`
-
-  `;
-  }
-}
